WIP: See details 🌩
- yml -> yaml for consistency
- Migrate code to top level dir
  Above will help future mkdocs summary of code
- Add soft links for offical yml content to code input folder
- Add shell script for regenerating all soft links
- WIP: split code for PEP compliance - 1 class:1 file
- WIP: Normalize code to load powers/bestiary similarly
