+++ README.md
@@ -1,10 +1,10 @@
-# Welcome to Deck of Adventures, The Game.
+# Welcome to Deck of Adventures, The Game
 
 This repository establishes the version history of the Deck of Adventure tabletop
 roleplaying game. A magical world awaits, where Players come first. In Deck of
 Adventure, it’s your turn, your choice.
 
-Deck of Adventures is a collaborative roleplaying system, built with an open source 
+Deck of Adventures is a collaborative roleplaying system, built with an open source
 ethos. Please vist our
 [Development Guide](./docs/src/2_Development/) to learn more about this project. Or visit our
 [Mechanics Guide](./docs/src/1_Mechanics/01_PlayerGuide_Full.md) to learn more about how to play.
@@ -13,3 +13,10 @@ ethos. Please vist our
 
 *If the above link is expired, please 
 [submit an issue](https://github.com/DeckofAdventures/TheGame/issues/new?assignees=&labels=bug&template=bug_report.md&title=Expired%20Discord%20Link!).*
+
+## Developers
+
+In addition to game documentation, this repository is home to automation tools that
+help us rapidly iterate on designs while keeping user-friendly outputs.
+
+In contributions, please conform to `black` formatting.
