+++ automation/_input/04_Powers.yaml
@@ -1,0 +1,1 @@
+../../docs/src/1_Mechanics/04_Powers.yaml
