+++ automation/_input/04_Powers_SAMPLE.yaml
