+++ automation/_input/05_Vulnerabilities.yaml
@@ -1,0 +1,1 @@
+../../docs/src/1_Mechanics/05_Vulnerabilities.yaml
