+++ automation/_input/06_Bestiary.yaml
@@ -1,0 +1,1 @@
+../../docs/src/1_Mechanics/06_Bestiary.yaml
