+++ automation/_input/06_Bestiary_SAMPLE.yaml
