+++ automation/pdf/__init__.py
@@ -1,0 +1,1 @@
+from .split_premades import *
