+++ automation/utils/logger.py
@@ -1,0 +1,27 @@
+"""This logger is adapted from the DataJoint Python API logger"""
+import logging
+import os
+import sys
+
+logger = logging.getLogger(__name__.split(".")[0])
+
+log_level = os.getenv("LOG_LEVEL", "INFO").upper()
+
+log_format = logging.Formatter("[%(asctime)s][%(levelname)s]: %(message)s")
+
+stream_handler = logging.StreamHandler()  # default handler
+stream_handler.setFormatter(log_format)
+
+logger.setLevel(level=log_level)
+logger.handlers = [stream_handler]
+
+
+def excepthook(exc_type, exc_value, exc_traceback):
+    if issubclass(exc_type, KeyboardInterrupt):
+        sys.__excepthook__(exc_type, exc_value, exc_traceback)
+        return
+
+    logger.error("Uncaught exception", exc_info=(exc_type, exc_value, exc_traceback))
+
+
+sys.excepthook = excepthook
