+++ docs/src/1_Mechanics/04_Powers.md
@@ -1,5 +1,4 @@
-<!-- DEVELOPERS: Please edit corresponding yml in 3_Automation -->
-
+<!-- DEVELOPERS: Please edit corresponding yml -->
 
 # Combat
