+++ docs/src/1_Mechanics/05_Vulnerabilities.md
@@ -1,4 +1,4 @@
-<!-- DEVELOPERS: Please edit corresponding yml in 3_Automation -->
+<!-- DEVELOPERS: Please edit corresponding yml -->
 
 
 # Combat
