+++ docs/src/1_Mechanics/05_Vulnerabilities.yaml
