+++ docs/src/1_Mechanics/06_Bestiary.yaml
