--- docs/src/3_Automation/README.md
@@ -1,12 +1,0 @@
-# Automation
-
-This folder contains tools to generate DofA materials. 
-
-For developers: please conform to `black` formatting.
-
-## Requirements
-
-To install requirements for existing scripts...
-```
-pip install -r requirements.txt
-```
