+++ docs/src/index.md
@@ -3,18 +3,18 @@ hide:
   - feedback
 ---
 
-# Welcome to Deck of Adventures, The Game.
+# Welcome to Deck of Adventures, The Game
 
-This repository establishes the version history of the Deck of Adventures tabletop
+This website establishes latest rules for the Deck of Adventures tabletop
 roleplaying game. A magical world awaits, where Players come first. In Deck of
 Adventures, it’s your turn, your choice.
 
-Deck of Adventures is a collaborative roleplaying system, built with an open source 
+Deck of Adventures is a collaborative roleplaying system, built with an open source
 ethos. Please vist our
 [Development Guide](/2_Development/01_Community/) to learn more about this project. Or visit our
 [Mechanics Guide](1_Mechanics/01_PlayerGuide_Full.md) to learn more about how to play.
 
 ## Join our [Discord server](https://discord.gg/dk6RfWgPHF)!
 
-*If the above link is expired, please 
+*If the above link is expired, please
 [submit an issue](https://github.com/DeckofAdventures/TheGame/issues/new?assignees=&labels=bug&template=bug_report.md&title=Expired%20Discord%20Link!).*
